# ApiGateway/app/openapi.yaml
openapi: 3.0.0
info:
  title: Hệ thống Quản lý Thanh toán Học phí (Microservices SOA)
  version: 1.0.0
  description: |
    Tài liệu API cho hệ thống Microservices bao gồm Auth, Customer Management, Payment Processor và Tuition Service.
    
    Tất cả các request **ngoại trừ /auth/login và /auth/logout** đều yêu cầu xác thực bằng **JWT Cookie** có tên là `jwt_token`. Gateway sẽ tự động trích xuất `customerId` từ JWT và thêm vào header `X-Customer-Id` trước khi gửi đến các dịch vụ backend.
servers:
  - url: http://localhost:8080 # Địa chỉ mặc định của API Gateway
security:
  - jwtCookie: [] # Áp dụng bảo mật cho các endpoint không phải là login/logout
...
paths:
...
  /payment/info:
    get:
      tags:
        - Payment Processor Service
      summary: Lấy lịch sử giao dịch thanh toán của khách hàng.
      operationId: getPaymentHistory
      responses:
        '200':
          description: Lịch sử giao dịch được truy xuất thành công.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/GenericResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/PaymentTransaction'
              example:
                success: true
                message: Payment history retrieved successfully.
                data:
                  - paymentId: 101
                    customerId: 1
                    tuitionId: 1
                    amount: 2800000
                    paymentDate: 2025-10-31T10:30:00Z
                  - paymentId: 102
                    customerId: 1
                    tuitionId: 3
                    amount: 1200000
                    paymentDate: 2025-10-25T09:15:00Z
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Không tìm thấy giao dịch nào cho khách hàng này.
          $ref: '#/components/responses/NotFoundGeneric'
        '500':
          $ref: '#/components/responses/InternalErrorGeneric'
  /payment/initiate:
    post:
      tags:
        - Payment Processor Service
      summary: Khởi tạo thanh toán và yêu cầu OTP (Bước 1).
      operationId: initiatePayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentInitiateRequest'
      responses:
        '200':
          description: OTP đã được gửi thành công.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/GenericResponse'
                  - type: object
                    properties:
                      data:
                        type: string
                        description: Thông báo hướng dẫn.
              example:
                success: true
                message: Payment initiation successful.
                data: OTP has been generated and sent to your email. Please use /confirm to finalize payment.
        '400':
          description: Lỗi nghiệp vụ ('Tuition ID already paid', 'Invalid request tuitionId is required.').
          $ref: '#/components/responses/BadRequestGeneric'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalErrorGeneric'
  /payment/confirm:
    post:
      tags:
        - Payment Processor Service
      summary: Xác nhận thanh toán bằng OTP (Bước 2).
      operationId: confirmPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentConfirmationRequest'
      responses:
        '200':
          description: Thanh toán thành công, trả về chi tiết giao dịch.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/GenericResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PaymentTransaction'
              example:
                success: true
                message: Payment confirmed and processed successfully.
                data:
                  paymentId: 101
                  customerId: 1
                  tuitionId: 1
                  amount: 2800000
                  paymentDate: 2025-10-31T10:35:00Z
        '400':
          description: Lỗi nghiệp vụ (ví dụ 'Invalid or expired OTP', 'Insufficient balance', 'Invalid request').
          $ref: '#/components/responses/BadRequestGeneric'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          description: Lỗi hệ thống nội bộ hoặc lỗi không nhất quán trạng thái (sau khi compensation thất bại).
          $ref: '#/components/responses/InternalErrorGeneric'
  /payment/resend:
    post:
      tags:
        - Payment Processor Service
      summary: Gửi lại OTP (Overwrites old OTP).
      operationId: resendOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentInitiateRequest'
      responses:
        '200':
          description: OTP mới đã được gửi thành công.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/GenericResponse'
                  - type: object
                    properties:
                      data:
                        type: string
                        description: Thông báo hướng dẫn.
              example:
                success: true
                message: Resending OTP successful.
                data: New OTP has been generated and sent to your email.
        '400':
          description: Lỗi nghiệp vụ (ví dụ 'Tuition ID already paid. Resend is unnecessary.').
          $ref: '#/components/responses/BadRequestGeneric'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalErrorGeneric'
components:
  schemas:
...
    GenericResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Cho biết request có thành công hay không.
          example: true
        message:
          type: string
          description: Thông báo mô tả kết quả.
          example: Current balance retrieved successfully.
        data:
          description: Dữ liệu trả về (kiểu dữ liệu cụ thể phụ thuộc vào endpoint).
          nullable: true
          example: null
...
    PaymentInitiateRequest:
      type: object
      description: Request body để khởi tạo thanh toán (yêu cầu OTP) hoặc gửi lại OTP.
      properties:
        tuitionId:
          type: integer
          format: int64
          example: 1
      required:
        - tuitionId
    PaymentConfirmationRequest:
      type: object
      description: Request body để xác nhận thanh toán bằng OTP.
      properties:
        tuitionId:
          type: integer
          format: int64
          example: 1
        otpCode:
          type: string
          example: "123456"
      required:
        - tuitionId
        - otpCode
    PaymentTransaction:
      type: object
      description: Hồ sơ giao dịch thanh toán đã thành công.
      properties:
        paymentId:
          type: integer
          format: int64
          example: 101
        customerId:
          type: integer
          format: int64
          example: 1
        tuitionId:
          type: integer
          format: int64
          example: 1
        amount:
          type: integer
          format: int32
          example: 2800000
        paymentDate:
          type: string
          format: date-time
          example: 2025-10-31T10:30:00Z
  securitySchemes:
    jwtCookie:
      type: apiKey
      in: cookie
      name: jwt_token
      description: JWT token được trả về sau khi đăng nhập thành công. (HttpOnly Cookie)
  responses:
    Unauthorized:
      description: Yêu cầu xác thực. JWT Cookie không được tìm thấy hoặc không hợp lệ/hết hạn.
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: Authentication required. JWT Cookie not found.
    BadRequestGeneric:
      description: Lỗi nghiệp vụ hoặc lỗi dữ liệu (400 Bad Request).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericResponse'
          example:
            success: false
            message: Insufficient balance to perform the debit. Required 2800000, available 1000000.
            data: null
    NotFoundGeneric:
      description: Không tìm thấy tài nguyên (404 Not Found).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericResponse'
          example:
            success: false
            message: Customer with ID 999 not found.
            data: null
...
    InternalErrorGeneric:
      description: Lỗi hệ thống nội bộ (500 Internal Server Error).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericResponse'
          example:
            success: false
            message: An unexpected error occurred on the server. Please contact support.
            data: null