# payment-processor-service/src/main/java/com/example/payment_processor_service/dto/PaymentInitiateRequest.java
package com.example.payment_processor_service.dto;

import lombok.Data;
import lombok.NoArgsConstructor;
import lombok.AllArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class PaymentInitiateRequest {
    private Long tuitionId;
}